move feature to feature folder
