+++ lib/home_page/feature/components/app_bar.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:unicons/unicons.dart';
 
-import '../../styles/styles.dart';
+import '../../../styles/styles.dart';
 import 'date_picker.dart';
 
 class HomePageAppBar extends StatelessWidget {
