+++ lib/home_page/home_page.dart
@@ -1,1 +1,1 @@
-export 'home_page_view.dart';
+export 'feature/home_page_view.dart';
